Struct aws_sdk_rekognition::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for RekognitionService.

This client allows ergonomic access to a RekognitionService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn compare_faces(&self) -> CompareFaces<C, M, R>

pub fn create_collection(&self) -> CreateCollection<C, M, R>

pub fn create_project(&self) -> CreateProject<C, M, R>

pub fn create_project_version(&self) -> CreateProjectVersion<C, M, R>

pub fn create_stream_processor(&self) -> CreateStreamProcessor<C, M, R>

pub fn delete_collection(&self) -> DeleteCollection<C, M, R>

pub fn delete_faces(&self) -> DeleteFaces<C, M, R>

pub fn delete_project(&self) -> DeleteProject<C, M, R>

pub fn delete_project_version(&self) -> DeleteProjectVersion<C, M, R>

pub fn delete_stream_processor(&self) -> DeleteStreamProcessor<C, M, R>

pub fn describe_collection(&self) -> DescribeCollection<C, M, R>

pub fn describe_projects(&self) -> DescribeProjects<C, M, R>

pub fn describe_project_versions(&self) -> DescribeProjectVersions<C, M, R>

pub fn describe_stream_processor(&self) -> DescribeStreamProcessor<C, M, R>

pub fn detect_custom_labels(&self) -> DetectCustomLabels<C, M, R>

pub fn detect_faces(&self) -> DetectFaces<C, M, R>

pub fn detect_labels(&self) -> DetectLabels<C, M, R>

pub fn detect_moderation_labels(&self) -> DetectModerationLabels<C, M, R>

pub fn detect_protective_equipment(&self) -> DetectProtectiveEquipment<C, M, R>

pub fn detect_text(&self) -> DetectText<C, M, R>

pub fn get_celebrity_info(&self) -> GetCelebrityInfo<C, M, R>

pub fn get_celebrity_recognition(&self) -> GetCelebrityRecognition<C, M, R>

pub fn get_content_moderation(&self) -> GetContentModeration<C, M, R>

pub fn get_face_detection(&self) -> GetFaceDetection<C, M, R>

pub fn get_face_search(&self) -> GetFaceSearch<C, M, R>

pub fn get_label_detection(&self) -> GetLabelDetection<C, M, R>

pub fn get_person_tracking(&self) -> GetPersonTracking<C, M, R>

pub fn get_segment_detection(&self) -> GetSegmentDetection<C, M, R>

pub fn get_text_detection(&self) -> GetTextDetection<C, M, R>

pub fn index_faces(&self) -> IndexFaces<C, M, R>

pub fn list_collections(&self) -> ListCollections<C, M, R>

pub fn list_faces(&self) -> ListFaces<C, M, R>

pub fn list_stream_processors(&self) -> ListStreamProcessors<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn recognize_celebrities(&self) -> RecognizeCelebrities<C, M, R>

pub fn search_faces(&self) -> SearchFaces<C, M, R>

pub fn search_faces_by_image(&self) -> SearchFacesByImage<C, M, R>

pub fn start_celebrity_recognition(&self) -> StartCelebrityRecognition<C, M, R>

pub fn start_content_moderation(&self) -> StartContentModeration<C, M, R>

pub fn start_face_detection(&self) -> StartFaceDetection<C, M, R>

pub fn start_face_search(&self) -> StartFaceSearch<C, M, R>

pub fn start_label_detection(&self) -> StartLabelDetection<C, M, R>

pub fn start_person_tracking(&self) -> StartPersonTracking<C, M, R>

pub fn start_project_version(&self) -> StartProjectVersion<C, M, R>

pub fn start_segment_detection(&self) -> StartSegmentDetection<C, M, R>

pub fn start_stream_processor(&self) -> StartStreamProcessor<C, M, R>

pub fn start_text_detection(&self) -> StartTextDetection<C, M, R>

pub fn stop_project_version(&self) -> StopProjectVersion<C, M, R>

pub fn stop_stream_processor(&self) -> StopStreamProcessor<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self