Module aws_sdk_rekognition::error

Modules

access_denied_exception

See AccessDeniedException

human_loop_quota_exceeded_exception

See HumanLoopQuotaExceededException

idempotent_parameter_mismatch_exception

See IdempotentParameterMismatchException

image_too_large_exception

See ImageTooLargeException

internal_server_error

See InternalServerError

invalid_image_format_exception

See InvalidImageFormatException

invalid_pagination_token_exception

See InvalidPaginationTokenException

invalid_parameter_exception

See InvalidParameterException

invalid_s3_object_exception

See InvalidS3ObjectException

limit_exceeded_exception

See LimitExceededException

provisioned_throughput_exceeded_exception

See ProvisionedThroughputExceededException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_in_use_exception

See ResourceInUseException

resource_not_found_exception

See ResourceNotFoundException

resource_not_ready_exception

See ResourceNotReadyException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

video_too_large_exception

See VideoTooLargeException

Structs

AccessDeniedException

You are not authorized to perform the action.

CompareFacesError

CreateCollectionError

CreateProjectError

CreateProjectVersionError

CreateStreamProcessorError

DeleteCollectionError

DeleteFacesError

DeleteProjectError

DeleteProjectVersionError

DeleteStreamProcessorError

DescribeCollectionError

DescribeProjectVersionsError

DescribeProjectsError

DescribeStreamProcessorError

DetectCustomLabelsError

DetectFacesError

DetectLabelsError

DetectModerationLabelsError

DetectProtectiveEquipmentError

DetectTextError

GetCelebrityInfoError

GetCelebrityRecognitionError

GetContentModerationError

GetFaceDetectionError

GetFaceSearchError

GetLabelDetectionError

GetPersonTrackingError

GetSegmentDetectionError

GetTextDetectionError

HumanLoopQuotaExceededException

The number of in-progress human reviews you have has exceeded the number allowed.

IdempotentParameterMismatchException

A ClientRequestToken input parameter was reused with an operation, but at least one of the other input

ImageTooLargeException

The input image size exceeds the allowed limit. If you are calling

IndexFacesError

InternalServerError

Amazon Rekognition experienced a service issue. Try your call again.

InvalidImageFormatException

The provided image format is not supported.

InvalidPaginationTokenException

Pagination token in the request is not valid.

InvalidParameterException

Input parameter violated a constraint. Validate your parameter before calling the API

InvalidS3ObjectException

Amazon Rekognition is unable to access the S3 object specified in the request.

LimitExceededException

An Amazon Rekognition service limit was exceeded. For example, if you start too many Amazon Rekognition Video jobs concurrently, calls to start operations

ListCollectionsError

ListFacesError

ListStreamProcessorsError

ListTagsForResourceError

ProvisionedThroughputExceededException

The number of requests exceeded your throughput limit. If you want to increase this

RecognizeCelebritiesError

ResourceAlreadyExistsException

A resource with the specified ID already exists.

ResourceInUseException

The specified resource is already being used.

ResourceNotFoundException

The resource specified in the request cannot be found.

ResourceNotReadyException

The requested resource isn't ready. For example,

SearchFacesByImageError

SearchFacesError

ServiceQuotaExceededException

StartCelebrityRecognitionError

StartContentModerationError

StartFaceDetectionError

StartFaceSearchError

StartLabelDetectionError

StartPersonTrackingError

StartProjectVersionError

StartSegmentDetectionError

StartStreamProcessorError

StartTextDetectionError

StopProjectVersionError

StopStreamProcessorError

TagResourceError

ThrottlingException

Amazon Rekognition is temporarily unable to process the request. Try your call again.

UntagResourceError

VideoTooLargeException

The file size or duration of the supplied media is too large. The maximum file size is 10GB.

Enums

CompareFacesErrorKind

CreateCollectionErrorKind

CreateProjectErrorKind

CreateProjectVersionErrorKind

CreateStreamProcessorErrorKind

DeleteCollectionErrorKind

DeleteFacesErrorKind

DeleteProjectErrorKind

DeleteProjectVersionErrorKind

DeleteStreamProcessorErrorKind

DescribeCollectionErrorKind

DescribeProjectVersionsErrorKind

DescribeProjectsErrorKind

DescribeStreamProcessorErrorKind

DetectCustomLabelsErrorKind

DetectFacesErrorKind

DetectLabelsErrorKind

DetectModerationLabelsErrorKind

DetectProtectiveEquipmentErrorKind

DetectTextErrorKind

GetCelebrityInfoErrorKind

GetCelebrityRecognitionErrorKind

GetContentModerationErrorKind

GetFaceDetectionErrorKind

GetFaceSearchErrorKind

GetLabelDetectionErrorKind

GetPersonTrackingErrorKind

GetSegmentDetectionErrorKind

GetTextDetectionErrorKind

IndexFacesErrorKind

ListCollectionsErrorKind

ListFacesErrorKind

ListStreamProcessorsErrorKind

ListTagsForResourceErrorKind

RecognizeCelebritiesErrorKind

SearchFacesByImageErrorKind

SearchFacesErrorKind

StartCelebrityRecognitionErrorKind

StartContentModerationErrorKind

StartFaceDetectionErrorKind

StartFaceSearchErrorKind

StartLabelDetectionErrorKind

StartPersonTrackingErrorKind

StartProjectVersionErrorKind

StartSegmentDetectionErrorKind

StartStreamProcessorErrorKind

StartTextDetectionErrorKind

StopProjectVersionErrorKind

StopStreamProcessorErrorKind

TagResourceErrorKind

UntagResourceErrorKind