Module aws_sdk_rekognition::model

Modules

age_range

See AgeRange

asset

See Asset

audio_metadata

See AudioMetadata

beard

See Beard

black_frame

See BlackFrame

bounding_box

See BoundingBox

celebrity

See Celebrity

celebrity_detail

See CelebrityDetail

celebrity_recognition

See CelebrityRecognition

compare_faces_match

See CompareFacesMatch

compared_face

See ComparedFace

compared_source_image_face

See ComparedSourceImageFace

content_moderation_detection

See ContentModerationDetection

covers_body_part

See CoversBodyPart

custom_label

See CustomLabel

detect_text_filters

See DetectTextFilters

detection_filter

See DetectionFilter

emotion

See Emotion

equipment_detection

See EquipmentDetection

evaluation_result

See EvaluationResult

eye_open

See EyeOpen

eyeglasses

See Eyeglasses

face

See Face

face_detail

See FaceDetail

face_detection

See FaceDetection

face_match

See FaceMatch

face_record

See FaceRecord

face_search_settings

See FaceSearchSettings

gender

See Gender

geometry

See Geometry

ground_truth_manifest

See GroundTruthManifest

human_loop_activation_output

See HumanLoopActivationOutput

human_loop_config

See HumanLoopConfig

human_loop_data_attributes

See HumanLoopDataAttributes

image

See Image

image_quality

See ImageQuality

instance

See Instance

kinesis_data_stream

See KinesisDataStream

kinesis_video_stream

See KinesisVideoStream

known_gender

See KnownGender

label

See Label

label_detection

See LabelDetection

landmark

See Landmark

moderation_label

See ModerationLabel

mouth_open

See MouthOpen

mustache

See Mustache

notification_channel

See NotificationChannel

output_config

See OutputConfig

parent

See Parent

person_detail

See PersonDetail

person_detection

See PersonDetection

person_match

See PersonMatch

point

See Point

pose

See Pose

project_description

See ProjectDescription

project_version_description

See ProjectVersionDescription

protective_equipment_body_part

See ProtectiveEquipmentBodyPart

protective_equipment_person

See ProtectiveEquipmentPerson

protective_equipment_summarization_attributes

See ProtectiveEquipmentSummarizationAttributes

protective_equipment_summary

See ProtectiveEquipmentSummary

region_of_interest

See RegionOfInterest

s3_object

See S3Object

segment_detection

See SegmentDetection

segment_type_info

See SegmentTypeInfo

shot_segment

See ShotSegment

smile

See Smile

start_segment_detection_filters

See StartSegmentDetectionFilters

start_shot_detection_filter

See StartShotDetectionFilter

start_technical_cue_detection_filter

See StartTechnicalCueDetectionFilter

start_text_detection_filters

See StartTextDetectionFilters

stream_processor

See StreamProcessor

stream_processor_input

See StreamProcessorInput

stream_processor_output

See StreamProcessorOutput

stream_processor_settings

See StreamProcessorSettings

summary

See Summary

sunglasses

See Sunglasses

technical_cue_segment

See TechnicalCueSegment

testing_data

See TestingData

testing_data_result

See TestingDataResult

text_detection

See TextDetection

text_detection_result

See TextDetectionResult

training_data

See TrainingData

training_data_result

See TrainingDataResult

unindexed_face

See UnindexedFace

validation_data

See ValidationData

video

See Video

video_metadata

See VideoMetadata

Structs

AgeRange

Structure containing the estimated age range, in years, for a face.

Asset

Assets are the images that you use to train and evaluate a model version.

AudioMetadata

Metadata information about an audio stream. An array of AudioMetadata objects

Beard

Indicates whether or not the face has a beard, and the confidence level in the

BlackFrame

BoundingBox

Identifies the bounding box around the label, face, text or personal protective equipment.

Celebrity

Provides information about a celebrity recognized by the RecognizeCelebrities operation.

CelebrityDetail

Information about a recognized celebrity.

CelebrityRecognition

Information about a detected celebrity and the time the celebrity was detected in a stored video.

CompareFacesMatch

Provides information about a face in a target image that matches the source image face

ComparedFace

Provides face metadata for target image faces that are analyzed by

ComparedSourceImageFace

Type that describes the face Amazon Rekognition chose to compare with the faces in the target.

ContentModerationDetection

Information about an inappropriate, unwanted, or offensive content label detection in a stored video.

CoversBodyPart

Information about an item of Personal Protective Equipment covering a corresponding body part. For more

CustomLabel

A custom label detected in an image by a call to DetectCustomLabels.

DetectTextFilters

A set of optional parameters that you can use to set the criteria that the text must meet to be included in your response.

DetectionFilter

A set of parameters that allow you to filter out certain results from your returned results.

Emotion

The emotions that appear to be expressed on the face, and the confidence level in the determination.

EquipmentDetection

Information about an item of Personal Protective Equipment (PPE) detected by

EvaluationResult

The evaluation results for the training of a model.

EyeOpen

Indicates whether or not the eyes on the face are open, and the confidence level in the

Eyeglasses

Indicates whether or not the face is wearing eye glasses, and the confidence level in

Face

Describes the face properties such as the bounding box, face ID, image ID of the input

FaceDetail

Structure containing attributes of the face that the algorithm detected.

FaceDetection

Information about a face detected in a video analysis request and the time the face was detected in the video.

FaceMatch

Provides face metadata. In addition, it also provides the confidence in the match of

FaceRecord

Object containing both the face metadata (stored in the backend database), and facial

FaceSearchSettings

Input face recognition parameters for an Amazon Rekognition stream processor. FaceRecognitionSettings is a request

Gender

The predicted gender of a detected face.

Geometry

Information about where an object (DetectCustomLabels) or text (DetectText) is located on

GroundTruthManifest

The S3 bucket that contains an Amazon Sagemaker Ground Truth format manifest file.

HumanLoopActivationOutput

Shows the results of the human in the loop evaluation. If there is no HumanLoopArn, the input did

HumanLoopConfig

Sets up the flow definition the image will be sent to if one of the conditions is met.

HumanLoopDataAttributes

Allows you to set attributes of the image. Currently, you can declare an image as free of

Image

Provides the input image either as bytes or an S3 object.

ImageQuality

Identifies face image brightness and sharpness.

Instance

An instance of a label returned by Amazon Rekognition Image (DetectLabels)

KinesisDataStream

The Kinesis data stream Amazon Rekognition to which the analysis results of a Amazon Rekognition stream processor are streamed. For more information, see

KinesisVideoStream

Kinesis video stream stream that provides the source streaming video for a Amazon Rekognition Video stream processor. For more information, see

KnownGender

The known gender identity for the celebrity that matches the provided ID.

Label

Structure containing details about the detected label, including the name, detected instances, parent labels, and level of

LabelDetection

Information about a label detected in a video analysis request and the time the label was detected in the video.

Landmark

Indicates the location of the landmark on the face.

ModerationLabel

Provides information about a single type of inappropriate, unwanted, or offensive content found in an image or video. Each type of

MouthOpen

Indicates whether or not the mouth on the face is open, and the confidence level in the

Mustache

Indicates whether or not the face has a mustache, and the confidence level in the

NotificationChannel

The Amazon Simple Notification Service topic to which Amazon Rekognition publishes the completion status of a video analysis operation. For more information, see

OutputConfig

The S3 bucket and folder location where training output is placed.

Parent

A parent label for a label. A label can have 0, 1, or more parents.

PersonDetail

Details about a person detected in a video analysis request.

PersonDetection

Details and path tracking information for a single time a person's path is tracked in a video.

PersonMatch

Information about a person whose face matches a face(s) in an Amazon Rekognition collection.

Point

The X and Y coordinates of a point on an image. The X and Y values returned are ratios

Pose

Indicates the pose of the face as determined by its pitch, roll, and yaw.

ProjectDescription

A description of a Amazon Rekognition Custom Labels project.

ProjectVersionDescription

The description of a version of a model.

ProtectiveEquipmentBodyPart

Information about a body part detected by DetectProtectiveEquipment that contains PPE.

ProtectiveEquipmentPerson

A person detected by a call to DetectProtectiveEquipment. The API returns

ProtectiveEquipmentSummarizationAttributes

Specifies summary attributes to return from a call to DetectProtectiveEquipment.

ProtectiveEquipmentSummary

Summary information for required items of personal protective equipment (PPE) detected on persons by a call

RegionOfInterest

Specifies a location within the frame that Rekognition checks for text. Uses a BoundingBox

S3Object

Provides the S3 bucket name and object name.

SegmentDetection

A technical cue or shot detection segment detected in a video. An array

SegmentTypeInfo

Information about the type of a segment requested in a call to StartSegmentDetection.

ShotSegment

Information about a shot detection segment detected in a video. For more information,

Smile

Indicates whether or not the face is smiling, and the confidence level in the

StartSegmentDetectionFilters

Filters applied to the technical cue or shot detection segments.

StartShotDetectionFilter

Filters for the shot detection segments returned by GetSegmentDetection.

StartTechnicalCueDetectionFilter

Filters for the technical segments returned by GetSegmentDetection. For more information,

StartTextDetectionFilters

Set of optional parameters that let you set the criteria text must meet to be included in your response.

StreamProcessor

An object that recognizes faces in a streaming video. An Amazon Rekognition stream processor is created by a call to CreateStreamProcessor. The request

StreamProcessorInput

Information about the source streaming video.

StreamProcessorOutput

Information about the Amazon Kinesis Data Streams stream to which a Amazon Rekognition Video stream processor streams the results of a video analysis. For more

StreamProcessorSettings

Input parameters used to recognize faces in a streaming video analyzed by a Amazon Rekognition stream processor.

Summary

The S3 bucket that contains the training summary. The training summary includes

Sunglasses

Indicates whether or not the face is wearing sunglasses, and the confidence level in

TechnicalCueSegment

Information about a technical cue segment. For more information, see SegmentDetection.

TestingData

The dataset used for testing. Optionally, if AutoCreate is set, Amazon Rekognition Custom Labels creates a

TestingDataResult

Sagemaker Groundtruth format manifest files for the input, output and validation datasets that are used and created during testing.

TextDetection

Information about a word or line of text detected by DetectText.

TextDetectionResult

Information about text detected in a video. Incudes the detected text,

TrainingData

The dataset used for training.

TrainingDataResult

Sagemaker Groundtruth format manifest files for the input, output and validation datasets that are used and created during testing.

UnindexedFace

A face that IndexFaces detected, but didn't index. Use the

ValidationData

Contains the Amazon S3 bucket location of the validation data for a model training job.

Video

Video file stored in an Amazon S3 bucket. Amazon Rekognition video start operations such as StartLabelDetection use Video to

VideoMetadata

Information about a video that Amazon Rekognition analyzed. Videometadata is returned in

Enums

Attribute

BodyPart

CelebrityRecognitionSortBy

ContentClassifier

ContentModerationSortBy

EmotionName

NOTE: EmotionName::Unknown has been renamed to ::UnknownValue.

FaceAttributes

FaceSearchSortBy

GenderType

KnownGenderType

A list of enum string of possible gender values that Celebrity returns.

LabelDetectionSortBy

LandmarkType

OrientationCorrection

PersonTrackingSortBy

ProjectStatus

ProjectVersionStatus

ProtectiveEquipmentType

QualityFilter

Reason

SegmentType

StreamProcessorStatus

TechnicalCueType

TextTypes

VideoColorRange

VideoJobStatus