Module aws_sdk_rekognition::operation

Structs

CompareFaces

Compares a face in the source input image with

CreateCollection

Creates a collection in an AWS Region. You can add faces to the collection using the

CreateProject

Creates a new Amazon Rekognition Custom Labels project. A project is a logical grouping of resources (images, Labels, models)

CreateProjectVersion

Creates a new version of a model and begins training.

CreateStreamProcessor

Creates an Amazon Rekognition stream processor that you can use to detect and recognize faces in a streaming video.

DeleteCollection

Deletes the specified collection. Note that this operation

DeleteFaces

Deletes faces from a collection. You specify a collection ID and an array of face IDs

DeleteProject

Deletes an Amazon Rekognition Custom Labels project. To delete a project you must first delete all models associated

DeleteProjectVersion

Deletes an Amazon Rekognition Custom Labels model.

DeleteStreamProcessor

Deletes the stream processor identified by Name. You assign the value for Name when you create the stream processor with

DescribeCollection

Describes the specified collection. You can use DescribeCollection to get

DescribeProjectVersions

Lists and describes the models in an Amazon Rekognition Custom Labels project. You

DescribeProjects

Lists and gets information about your Amazon Rekognition Custom Labels projects.

DescribeStreamProcessor

Provides information about a stream processor created by CreateStreamProcessor. You can get information about the input and output streams, the input parameters for the face recognition being performed,

DetectCustomLabels

Detects custom labels in a supplied image by using an Amazon Rekognition Custom Labels model.

DetectFaces

Detects faces within an image that is provided as input.

DetectLabels

Detects instances of real-world entities within an image (JPEG or PNG)

DetectModerationLabels

Detects unsafe content in a specified JPEG or PNG format image.

DetectProtectiveEquipment

Detects Personal Protective Equipment (PPE) worn by people detected in an image. Amazon Rekognition can detect the

DetectText

Detects text in the input image and converts it into machine-readable text.

GetCelebrityInfo

Gets the name and additional information about a celebrity based on their Amazon Rekognition ID.

GetCelebrityRecognition

Gets the celebrity recognition results for a Amazon Rekognition Video analysis started by

GetContentModeration

Gets the inappropriate, unwanted, or offensive content analysis results for a Amazon Rekognition Video analysis started by

GetFaceDetection

Gets face detection results for a Amazon Rekognition Video analysis started by StartFaceDetection.

GetFaceSearch

Gets the face search results for Amazon Rekognition Video face search started by

GetLabelDetection

Gets the label detection results of a Amazon Rekognition Video analysis started by StartLabelDetection.

GetPersonTracking

Gets the path tracking results of a Amazon Rekognition Video analysis started by StartPersonTracking.

GetSegmentDetection

Gets the segment detection results of a Amazon Rekognition Video analysis started by StartSegmentDetection.

GetTextDetection

Gets the text detection results of a Amazon Rekognition Video analysis started by StartTextDetection.

IndexFaces

Detects faces in the input image and adds them to the specified collection.

ListCollections

Returns list of collection IDs in your account.

ListFaces

Returns metadata for faces in the specified collection.

ListStreamProcessors

Gets a list of stream processors that you have created with CreateStreamProcessor.

ListTagsForResource

RecognizeCelebrities

Returns an array of celebrities recognized in the input image. For more information, see Recognizing Celebrities

SearchFaces

For a given input face ID, searches for matching faces in the collection the face

SearchFacesByImage

For a given input image, first detects the largest face in the image, and then searches

StartCelebrityRecognition

Starts asynchronous recognition of celebrities in a stored video.

StartContentModeration

Starts asynchronous detection of inappropriate, unwanted, or offensive content in a stored video. For a list of moderation labels in Amazon Rekognition, see

StartFaceDetection

Starts asynchronous detection of faces in a stored video.

StartFaceSearch

Starts the asynchronous search for faces in a collection that match the faces of persons detected in a stored video.

StartLabelDetection

Starts asynchronous detection of labels in a stored video.

StartPersonTracking

Starts the asynchronous tracking of a person's path in a stored video.

StartProjectVersion

Starts the running of the version of a model. Starting a model takes a while

StartSegmentDetection

Starts asynchronous detection of segment detection in a stored video.

StartStreamProcessor

Starts processing a stream processor. You create a stream processor by calling CreateStreamProcessor.

StartTextDetection

Starts asynchronous detection of text in a stored video.

StopProjectVersion

Stops a running model. The operation might take a while to complete. To

StopStreamProcessor

Stops a running stream processor that was created by CreateStreamProcessor.

TagResource

UntagResource