Struct aws_sdk_rekognition::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for RekognitionService.

This client allows ergonomic access to a RekognitionService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn compare_faces(&self) -> CompareFaces<C, M, R>

Constructs a fluent builder for the CompareFaces operation.

See CompareFaces for more information about the operation and its arguments.

pub fn create_collection(&self) -> CreateCollection<C, M, R>

Constructs a fluent builder for the CreateCollection operation.

See CreateCollection for more information about the operation and its arguments.

pub fn create_dataset(&self) -> CreateDataset<C, M, R>

Constructs a fluent builder for the CreateDataset operation.

See CreateDataset for more information about the operation and its arguments.

pub fn create_project(&self) -> CreateProject<C, M, R>

Constructs a fluent builder for the CreateProject operation.

See CreateProject for more information about the operation and its arguments.

pub fn create_project_version(&self) -> CreateProjectVersion<C, M, R>

Constructs a fluent builder for the CreateProjectVersion operation.

See CreateProjectVersion for more information about the operation and its arguments.

pub fn create_stream_processor(&self) -> CreateStreamProcessor<C, M, R>

Constructs a fluent builder for the CreateStreamProcessor operation.

See CreateStreamProcessor for more information about the operation and its arguments.

pub fn delete_collection(&self) -> DeleteCollection<C, M, R>

Constructs a fluent builder for the DeleteCollection operation.

See DeleteCollection for more information about the operation and its arguments.

pub fn delete_dataset(&self) -> DeleteDataset<C, M, R>

Constructs a fluent builder for the DeleteDataset operation.

See DeleteDataset for more information about the operation and its arguments.

pub fn delete_faces(&self) -> DeleteFaces<C, M, R>

Constructs a fluent builder for the DeleteFaces operation.

See DeleteFaces for more information about the operation and its arguments.

pub fn delete_project(&self) -> DeleteProject<C, M, R>

Constructs a fluent builder for the DeleteProject operation.

See DeleteProject for more information about the operation and its arguments.

pub fn delete_project_version(&self) -> DeleteProjectVersion<C, M, R>

Constructs a fluent builder for the DeleteProjectVersion operation.

See DeleteProjectVersion for more information about the operation and its arguments.

pub fn delete_stream_processor(&self) -> DeleteStreamProcessor<C, M, R>

Constructs a fluent builder for the DeleteStreamProcessor operation.

See DeleteStreamProcessor for more information about the operation and its arguments.

pub fn describe_collection(&self) -> DescribeCollection<C, M, R>

Constructs a fluent builder for the DescribeCollection operation.

See DescribeCollection for more information about the operation and its arguments.

pub fn describe_dataset(&self) -> DescribeDataset<C, M, R>

Constructs a fluent builder for the DescribeDataset operation.

See DescribeDataset for more information about the operation and its arguments.

pub fn describe_projects(&self) -> DescribeProjects<C, M, R>

Constructs a fluent builder for the DescribeProjects operation.

See DescribeProjects for more information about the operation and its arguments.

pub fn describe_project_versions(&self) -> DescribeProjectVersions<C, M, R>

Constructs a fluent builder for the DescribeProjectVersions operation.

See DescribeProjectVersions for more information about the operation and its arguments.

pub fn describe_stream_processor(&self) -> DescribeStreamProcessor<C, M, R>

Constructs a fluent builder for the DescribeStreamProcessor operation.

See DescribeStreamProcessor for more information about the operation and its arguments.

pub fn detect_custom_labels(&self) -> DetectCustomLabels<C, M, R>

Constructs a fluent builder for the DetectCustomLabels operation.

See DetectCustomLabels for more information about the operation and its arguments.

pub fn detect_faces(&self) -> DetectFaces<C, M, R>

Constructs a fluent builder for the DetectFaces operation.

See DetectFaces for more information about the operation and its arguments.

pub fn detect_labels(&self) -> DetectLabels<C, M, R>

Constructs a fluent builder for the DetectLabels operation.

See DetectLabels for more information about the operation and its arguments.

pub fn detect_moderation_labels(&self) -> DetectModerationLabels<C, M, R>

Constructs a fluent builder for the DetectModerationLabels operation.

See DetectModerationLabels for more information about the operation and its arguments.

pub fn detect_protective_equipment(&self) -> DetectProtectiveEquipment<C, M, R>

Constructs a fluent builder for the DetectProtectiveEquipment operation.

See DetectProtectiveEquipment for more information about the operation and its arguments.

pub fn detect_text(&self) -> DetectText<C, M, R>

Constructs a fluent builder for the DetectText operation.

See DetectText for more information about the operation and its arguments.

pub fn distribute_dataset_entries(&self) -> DistributeDatasetEntries<C, M, R>

Constructs a fluent builder for the DistributeDatasetEntries operation.

See DistributeDatasetEntries for more information about the operation and its arguments.

pub fn get_celebrity_info(&self) -> GetCelebrityInfo<C, M, R>

Constructs a fluent builder for the GetCelebrityInfo operation.

See GetCelebrityInfo for more information about the operation and its arguments.

pub fn get_celebrity_recognition(&self) -> GetCelebrityRecognition<C, M, R>

Constructs a fluent builder for the GetCelebrityRecognition operation.

See GetCelebrityRecognition for more information about the operation and its arguments.

pub fn get_content_moderation(&self) -> GetContentModeration<C, M, R>

Constructs a fluent builder for the GetContentModeration operation.

See GetContentModeration for more information about the operation and its arguments.

pub fn get_face_detection(&self) -> GetFaceDetection<C, M, R>

Constructs a fluent builder for the GetFaceDetection operation.

See GetFaceDetection for more information about the operation and its arguments.

pub fn get_face_search(&self) -> GetFaceSearch<C, M, R>

Constructs a fluent builder for the GetFaceSearch operation.

See GetFaceSearch for more information about the operation and its arguments.

pub fn get_label_detection(&self) -> GetLabelDetection<C, M, R>

Constructs a fluent builder for the GetLabelDetection operation.

See GetLabelDetection for more information about the operation and its arguments.

pub fn get_person_tracking(&self) -> GetPersonTracking<C, M, R>

Constructs a fluent builder for the GetPersonTracking operation.

See GetPersonTracking for more information about the operation and its arguments.

pub fn get_segment_detection(&self) -> GetSegmentDetection<C, M, R>

Constructs a fluent builder for the GetSegmentDetection operation.

See GetSegmentDetection for more information about the operation and its arguments.

pub fn get_text_detection(&self) -> GetTextDetection<C, M, R>

Constructs a fluent builder for the GetTextDetection operation.

See GetTextDetection for more information about the operation and its arguments.

pub fn index_faces(&self) -> IndexFaces<C, M, R>

Constructs a fluent builder for the IndexFaces operation.

See IndexFaces for more information about the operation and its arguments.

pub fn list_collections(&self) -> ListCollections<C, M, R>

Constructs a fluent builder for the ListCollections operation.

See ListCollections for more information about the operation and its arguments.

pub fn list_dataset_entries(&self) -> ListDatasetEntries<C, M, R>

Constructs a fluent builder for the ListDatasetEntries operation.

See ListDatasetEntries for more information about the operation and its arguments.

pub fn list_dataset_labels(&self) -> ListDatasetLabels<C, M, R>

Constructs a fluent builder for the ListDatasetLabels operation.

See ListDatasetLabels for more information about the operation and its arguments.

pub fn list_faces(&self) -> ListFaces<C, M, R>

Constructs a fluent builder for the ListFaces operation.

See ListFaces for more information about the operation and its arguments.

pub fn list_stream_processors(&self) -> ListStreamProcessors<C, M, R>

Constructs a fluent builder for the ListStreamProcessors operation.

See ListStreamProcessors for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn recognize_celebrities(&self) -> RecognizeCelebrities<C, M, R>

Constructs a fluent builder for the RecognizeCelebrities operation.

See RecognizeCelebrities for more information about the operation and its arguments.

pub fn search_faces(&self) -> SearchFaces<C, M, R>

Constructs a fluent builder for the SearchFaces operation.

See SearchFaces for more information about the operation and its arguments.

pub fn search_faces_by_image(&self) -> SearchFacesByImage<C, M, R>

Constructs a fluent builder for the SearchFacesByImage operation.

See SearchFacesByImage for more information about the operation and its arguments.

pub fn start_celebrity_recognition(&self) -> StartCelebrityRecognition<C, M, R>

Constructs a fluent builder for the StartCelebrityRecognition operation.

See StartCelebrityRecognition for more information about the operation and its arguments.

pub fn start_content_moderation(&self) -> StartContentModeration<C, M, R>

Constructs a fluent builder for the StartContentModeration operation.

See StartContentModeration for more information about the operation and its arguments.

pub fn start_face_detection(&self) -> StartFaceDetection<C, M, R>

Constructs a fluent builder for the StartFaceDetection operation.

See StartFaceDetection for more information about the operation and its arguments.

pub fn start_face_search(&self) -> StartFaceSearch<C, M, R>

Constructs a fluent builder for the StartFaceSearch operation.

See StartFaceSearch for more information about the operation and its arguments.

pub fn start_label_detection(&self) -> StartLabelDetection<C, M, R>

Constructs a fluent builder for the StartLabelDetection operation.

See StartLabelDetection for more information about the operation and its arguments.

pub fn start_person_tracking(&self) -> StartPersonTracking<C, M, R>

Constructs a fluent builder for the StartPersonTracking operation.

See StartPersonTracking for more information about the operation and its arguments.

pub fn start_project_version(&self) -> StartProjectVersion<C, M, R>

Constructs a fluent builder for the StartProjectVersion operation.

See StartProjectVersion for more information about the operation and its arguments.

pub fn start_segment_detection(&self) -> StartSegmentDetection<C, M, R>

Constructs a fluent builder for the StartSegmentDetection operation.

See StartSegmentDetection for more information about the operation and its arguments.

pub fn start_stream_processor(&self) -> StartStreamProcessor<C, M, R>

Constructs a fluent builder for the StartStreamProcessor operation.

See StartStreamProcessor for more information about the operation and its arguments.

pub fn start_text_detection(&self) -> StartTextDetection<C, M, R>

Constructs a fluent builder for the StartTextDetection operation.

See StartTextDetection for more information about the operation and its arguments.

pub fn stop_project_version(&self) -> StopProjectVersion<C, M, R>

Constructs a fluent builder for the StopProjectVersion operation.

See StopProjectVersion for more information about the operation and its arguments.

pub fn stop_stream_processor(&self) -> StopStreamProcessor<C, M, R>

Constructs a fluent builder for the StopStreamProcessor operation.

See StopStreamProcessor for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_dataset_entries(&self) -> UpdateDatasetEntries<C, M, R>

Constructs a fluent builder for the UpdateDatasetEntries operation.

See UpdateDatasetEntries for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.